#40-UserRepo определил пакет для перечислений, переименовал файл со скриптом, изменил тип для полей связанных с id, добавил путь для хранения image и определил поведение liquibase в файле application.properties

+++ src/main/java/ru/skypro/homework/dto/ad/Ad.java
@@ -6,9 +6,9 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class Ad {
-    private long author;
+    private int author;
     private String image;
-    private long pk;
+    private int pk;
     private int price;
     private String title;
 }
